feat: expose containers on port 80 pending upstream TLS proxy
Realised that terminating TLS at the load balancer (as I normally would)
using ACM and Route 53 would not be free tier usage. I would like to get
all resources within this project so it's reproducible and independent.

Therefore a cloudfront distribution which comes with an issued
certificate and autogenerated domain is my preferred option now.